move store to use folder
