+++ src/api/Common.ts
@@ -1,4 +1,4 @@
-import { useStore } from "@/store/store";
+import { useStore } from "@/use/store/store";
 import axios, { AxiosResponse, AxiosError } from "axios";
 import { AxiosInstance } from "axios";
