+++ src/components/settings/SecuritySection.vue
@@ -91,7 +91,7 @@
 
 <script lang="ts">
     import { ref, computed } from "vue";
-    import { store } from "@/store/store";
+    import { store } from "@/use/store/store";
     import EnterPasswordModal from "@/components/modals/EnterPasswordModal.vue";
     import UserManagement from "@/api/UserManagement";
     import GenericResponseHandler from "@/use/GenericResponseHandler";
