+++ src/main.ts
@@ -1,7 +1,7 @@
 import { createApp } from "vue";
 import App from "./App.vue";
 import Router from "./router";
-import { store } from "./store/store";
+import { store } from "./use/store/store";
 import "./assets/styles/index.css";
 
 const app = createApp(App);
