+++ src/router/index.ts
@@ -13,7 +13,7 @@ import WelcomePage from "../views/common/Welcome.vue";
 import AboutPage from "../views/common/About.vue";
 import { checkPrivilege } from "@/use/PermissionHelper";
 import { Role } from "@/entities/Role";
-import { useStore } from "@/store/store";
+import { useStore } from "@/use/store/store";
 
 const routerHistory = createWebHistory(process.env.BASE_URL);
 const suffix: string = " | UC4";
