+++ src/use/PermissionHelper.ts
@@ -1,5 +1,5 @@
 import { Role } from "@/entities/Role";
-import { useStore } from "@/store/store";
+import { useStore } from "@/use/store/store";
 
 export async function checkPrivilege(...roles: Role[]): Promise<{ authenticated: boolean; allowed: boolean }> {
     const store = useStore();
