+++ src/views/common/Login.vue
@@ -68,7 +68,7 @@
 </template>
 <script lang="ts">
     import Router from "@/router/";
-    import { useStore, store } from "../../store/store";
+    import { useStore, store } from "@/use/store/store";
     import { Role } from "../../entities/Role";
     import UserManagement from "@/api/UserManagement";
     import { ref, onMounted } from "vue";
